Navbar refactor
