--- client/src/components/Navbar/NavRight.styles.js
@@ -1,17 +1,0 @@
-import styled from 'styled-components';
-import { mobile } from '../../responsive';
-
-export const NavRightContainer = styled.div`
-    flex: 1;
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-    ${mobile({ flex: 2, justifyContent: 'center' })}
-`;
-
-export const MenuItem = styled.div`
-    font-size: 14px;
-    cursor: pointer;
-    margin-left: 25px;
-    ${mobile({ fontSize: '12px', marginLeft: '10px' })}
-`;
