+++ client/src/components/Navbar/__test__/Navbar.test.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import Navbar from '../Navbar';
 import '@testing-library/jest-dom';
+import { ShoppingCartOutlined } from '@material-ui/icons';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { fireEvent, screen } from '@testing-library/react';
-import { Language, Logo, SearchInput } from '../Navbar.styles';
+import { Language, Logo, SearchInput } from '../Navbar';
 import { renderWithProviders } from '../../../utils/test-utils';
 
 describe('Navbar', () => {
@@ -50,5 +51,31 @@ describe('Navbar', () => {
         const heading = screen.getByRole('heading', { name: 'Better Buys' });
         expect(heading).toBeInTheDocument();
         expect(heading).toHaveTextContent('Better Buys');
+    });
+
+    it('checks if clicking on the "SIGN UP" button in the Navbar component navigates the user to the signup page', () => {
+        renderWithProviders(
+            <Router>
+                <Navbar />
+            </Router>,
+        );
+        const signUpButton = screen.getByText('SIGN UP');
+        fireEvent.click(signUpButton);
+        expect(window.location.pathname).toBe('/signup');
+    });
+
+    it('checks if the "SIGN IN" text is properly rendered in the Navbar component', async () => {
+        renderWithProviders(
+            <Router>
+                <Navbar />
+            </Router>,
+        );
+        const signinElement = screen.getByText(/SIGN IN/i);
+        expect(signinElement).toBeInTheDocument();
+    });
+
+    it('checking that the shopping cart icon is rendered within the component', () => {
+        renderWithProviders(<ShoppingCartOutlined data-testid="cart-icon" />);
+        expect(screen.getByTestId('cart-icon')).toBeInTheDocument();
     });
 });
